requirements.txt
Django==3.2.14
djangorestframework==3.13.1
requests==2.27.1
# markdown==3.3.6
django-session-timeout==0.1.0
django-rest-swagger==2.2.0
django-extensions==3.1.5
django-webpack-loader==1.6.0
django-filter